Allow the name to be passed into add-scaffolding
The default declaration of $*NULL gets tucked into the block in the
declaration's assignment during composition. This is a problem because
when you add it to the scaffolding you are just adding a block rather
than a constant declaration, so it's impossible to retrieve it by name
in the compilation stage. This change makes the composer pass in the
name of the original declarable to the DependencyList.