automatically optimize SAST::Blocks/Stmts away
Instead of doing it in the compiler the composer now tries to reduce
every SAST::Stmts as much as possible except for those marked with
`$.auto-inline = False`. This means that Blocks in if statements won't
be replaced but blocks in the body of another block will be if they can.