Auto type cast the return values of coercer methods
with things like:

class Str {
    method Int { "1" }
}

Because "1".WHAT is Str, spit would go into an infinite loop trying to
recursively use the Int method to change the return value of the Int
method itself into an Int.

Fix this by forcibly type casting if you are declaring a coercer
method for a primitive type.