Introduce SAST::Force to take over from Blessed in stage 3+
Wrapping a node in a SAST::Blessed used to be the way to force its
type. The problem with that was it created extra nodes everywhere
which made the AST more complex. This led to things not being
optimized/composed the right way.

Solution: Instead of wrapping nodes, .^mixin them with a
role (SAST::Force) which simply overrides their .type and .itemize
methods.

Doing this led to a lot more stuff getting inlined which of course
exposed some bugs which have been fixed in this commit.