Change sub foo(...-->Type) to sub foo(...)-->Type
Since --> is a postfix we can't have it in our signature otherwise the
following is ambiguous:

sub foo($a = "foo"-->Int) { }

Is the sub returning an Int or is "foo" being type cast to an Int?

Putting it outside actually looks better anyway and fits in with
future changes to return type shorthands I've got coming up.