Rework command silencing after inlining
Use to set a $!silence attribute on SAST::Cmd when we find out
in stage3 that a command needs to be sent to /dev/null. But it only
sorta worked out because it depends on $*NULL but doesn't add it as a
dependency. This meant it had to rely on the scaffolding system which
turns out to unreliable.

Instead, just do the hard messy work to properly set the SAST::Cmd's
stdout to $*NULL in stage3.