Fix problems with $a && '{foo bar}' like statements
since $a && ... is compiled as ${a:+...} you have to be careful on how
you compile the RHS. It is a special case that behaves like it is in
"..." quotes except that you also need to escape '}'. I had to
introduce a new shell element type called 'Concat' to propagate
whether we're in parameter expansion or double quotes to its
sub-elements so they escape the right things.