introduce SAST::Noisy instead of $*NULL in scaffolding
/dev/null ($*NULL) often needed to be used by the compiler when it
decided that something was going to put stuff to stdout that
shouldn't. I tried to use scaffolding to get a late reference to it. This didn't
always work well because $*NULL is an option and therefore too complex to be
scaffolding.

Instead things that might need to be silenced are given the SAST::Noisy
role and a reference to $*NULL is put into them if a condition is met
during composition.