Fix $_ being undef in blocks in expressions with statement modifiers
in Something like:
{say $_} if $foo

$_ would be undefined because the {...} would set its outer scope as
normal and then would effectively ignore the scope that the if
statement tries to wrap around it.

Now whenever we have a statement modifier we recursively descend the
statement and fix up any blocks to have their outer scope be the
synthetic lexical scope the if statement modifier creates.