Fix complicated pipe bug
TIL: with, echo "foo" | echo $(cat)
The $(cat) will be evaluated in the pipe subshell. Previously I
thought that it was evaluated before the pipe was set up in the main
shell. This means that if there is an implicitly piped $self as an
argument to something that is being piped (like another method call)
the STDIN representing the implicit $self will get clobbered. So this
patch makes sure things won't get implicitly piped in those
situations.

TODO: check whether the problem can happen in ${...} commands (hint:
it can)