Calling rindex on a List comes with a warning
This was changed a few months ago in Raku.  Since `List`s are `Cool`, they get automatically stringified on `Str` methods, such as `substr` and `rindex`.  The latter nowadays warns as it usually indicative of a thinko in the developer's mind.  By adding a explicit `join`, the warning is prevented.  From a performance point of view, there is no difference, as the same `join` would have been done under the hood.